Issue #329 Clean up home component and reduce unecessary number of files
