--- src/components/home_page/home_page_component.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Content } from 'native-base';
-import { applicationStyles } from '../../application/styles';
-import { CopyrightComponent } from './copyright_component';
-import { PersonalizeComponent } from './personalize_component';
-import { MyToolsComponent } from './my_tools_component';
-import { History } from 'history';
-
-export interface HomePageProps {
-    readonly history: History;
-}
-
-export const HomePageComponent: React.StatelessComponent<HomePageProps> = (props: HomePageProps): JSX.Element => {
-    return (
-        <Content padder style={applicationStyles.body} >
-            <PersonalizeComponent history={props.history} />
-            <MyToolsComponent history={props.history} />
-            <CopyrightComponent />
-        </Content >
-    );
-};
