+++ src/components/main/main_page_switcher.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-native';
 import { WelcomeConnectedComponent } from '../welcome/welcome_connected_component';
-import { HomePageComponent } from '../home_page/home_page_component';
+import { HomeComponent } from '../home/home_component';
 import { HelpConnectedComponent } from '../help/help_connected_component';
 import { ExploreAllConnectedComponent } from '../explore/explore_all_connected_component';
 import { ExploreDetailConnectedComponent } from '../explore/explore_detail_connected_component';
@@ -16,7 +16,7 @@ import { Routes, routePathDefinition } from '../../application/routing';
 export const MainPageSwitcherComponent: React.StatelessComponent = (): JSX.Element => (
     <Switch>
         <Route exact path={routePathDefinition(Routes.Welcome)} component={WelcomeConnectedComponent} />
-        <Route exact path={routePathDefinition(Routes.Home)} component={HomePageComponent} />
+        <Route exact path={routePathDefinition(Routes.Home)} component={HomeComponent} />
         <Route exact path={routePathDefinition(Routes.Help)} component={HelpConnectedComponent} />
         <Route exact path={routePathDefinition(Routes.Questionnaire)} component={QuestionnaireConnectedComponent} />
         <Route exact path={routePathDefinition(Routes.Learn)} component={ExploreAllConnectedComponent} />
